pyproject.toml
[project]
name = "crtgen"
version = "0.1.0"
description = "Automating the creation of SecureCRT session files"
authors = [
    {name = "Dongju Yoo",email = "you1367@gmail.com"}
]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "rich (>=14.0.0,<15.0.0)",
    "pycryptodome (>=3.22.0,<4.0.0)"
]

[tool.poetry]
packages = [{include = "crtgen", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
